+++ tests/Http/Livewire/BaseTable.php
@@ -3,60 +3,12 @@
 namespace Rappasoft\LaravelLivewireTables\Tests\Http\Livewire;
 
 use Rappasoft\LaravelLivewireTables\DataTableComponent;
+use Rappasoft\LaravelLivewireTables\Tests\Http\Livewire\Traits\TestHelpers;
 
 abstract class BaseTable extends DataTableComponent
 {
+    use TestHelpers;
+
     public string $paginationTest = 'standard';
 
-    public function enableDetailedPagination(string $type = 'standard')
-    {
-        $this->setPerPageAccepted([1, 3, 5, 10, 15, 25, 50])->setPerPage(3);
-        $this->setPaginationMethod($type);
-        $this->setDisplayPaginationDetailsEnabled();
-
-    }
-
-    public function disableDetailedPagination(string $type = 'standard')
-    {
-        $this->setPerPageAccepted([1, 3, 5, 10, 15, 25, 50])->setPerPage(3);
-        $this->setPaginationMethod($type);
-        $this->setDisplayPaginationDetailsDisabled();
-    }
-
-    public function setPaginationTest(string $type)
-    {
-        $this->paginationTest = $type;
-    }
-
-    public function bootAll()
-    {
-        $view = view('livewire-tables::datatable');
-
-        $this->boot();
-        $this->bootedComponentUtilities();
-        $this->bootedHasFiltersCore();
-        $this->bootedWithColumns();
-        $this->bootedWithColumnSelect();
-        $this->booted();
-        $this->mountHasFiltersCore();
-        $this->mountComponentUtilities();
-        $this->mountWithSorting();
-        $this->renderAll($view);
-    }
-
-    public function renderAll($view = null)
-    {
-        if (is_null($view)) {
-            $view = view('livewire-tables::datatable');
-        }
-        $this->renderingWithColumns($view, $view->getData());
-        $this->renderingWithColumnSelect($view, $view->getData());
-        $this->renderingWithCustomisations($view, $view->getData());
-        $this->renderingWithData($view, $view->getData());
-        $this->renderingWithReordering($view, $view->getData());
-        $this->renderingWithPagination($view, $view->getData());
-        $this->render();
-
-        return $view;
-    }
 }
